Create the core library
