+++ packages/core/.babelrc
