+++ packages/core/.travis.yml
