+++ packages/core/package.json
@@ -1,21 +1,27 @@
 {
-  "name": "puzzle-lib",
+  "name": "core",
   "version": "1.0.0",
   "description": "A library to play the 15 puzzle game",
-  "main": "lib/library.js",
+  "main": "build/core.js",
   "scripts": {
-    "build": "webpack --env build"
+    "build": "webpack --env dev && webpack --env build"
   },
   "devDependencies": {
     "babel-cli": "~6.26.0",
     "babel-core": "~6.26.0",
     "babel-eslint": "~8.0.2",
+    "babel-jest": "~21.2.0",
     "babel-loader": "~7.1.2",
     "babel-plugin-add-module-exports": "~0.2.1",
     "babel-preset-es2015": "~6.24.1",
     "eslint": "~4.12.0",
     "eslint-loader": "~1.9.0",
+    "eslint-config-prettier": "~2.8.0",
+    "eslint-plugin-import": "~2.8.0",
+    "eslint-plugin-jest": "~21.3.2",
+    "eslint-plugin-prettier": "~2.3.1",
     "jest": "~21.2.1",
+    "prettier": "~1.8.2",
     "webpack": "~3.8.1",
     "yargs": "~10.0.3"
   }
