+++ packages/core/src/helper.js
@@ -2,6 +2,7 @@ export const associateTileToBackground = grid => {
     const tileSize = grid.length > 0 ? 100 / grid.length : 0; // 100%
 
     let tileToBg = {};
+
     grid.forEach((row, rowKey) => {
         row.forEach((value, colKey) => {
             tileToBg[value] = `${tileSize * colKey}% ${tileSize * rowKey}%`;
@@ -16,7 +17,7 @@ export const choiceInArray = (list, rand = Math.random()) => {
     }
     if (rand < 0 || rand >= 1) {
         throw Error(
-            `The value random number should be a number (0 <= rand < 1)`,
+            'The value random number should be a number (0 <= rand < 1)',
         );
     }
