+++ packages/core/src/helper.spec.js
@@ -28,6 +28,7 @@ describe('Helper', () => {
                 0: '75% 75%',
             };
             const association = Helper.associateTileToBackground(grid);
+
             expect(association).toEqual(expectedAssociation);
         });
 
@@ -35,6 +36,7 @@ describe('Helper', () => {
             const grid = [];
 
             const association = Helper.associateTileToBackground(grid);
+
             expect(association).toEqual({});
         });
     });
@@ -43,6 +45,7 @@ describe('Helper', () => {
         test('should return the first item', () => {
             const list = ['a', 'b', 'c'];
             const item = Helper.choiceInArray(list, 0);
+
             expect(item).toEqual(list[0]);
         });
 
@@ -50,6 +53,7 @@ describe('Helper', () => {
             const spy = jest.spyOn(Math, 'random');
 
             const list = ['a', 'b', 'c'];
+
             Helper.choiceInArray(list);
             expect(spy).toHaveBeenCalled();
 
@@ -59,13 +63,15 @@ describe('Helper', () => {
 
         test('should throw an error if the rand number is not between 0 and 1', () => {
             const list = ['a', 'b', 'c'];
+
             expect(() => Helper.choiceInArray(list, 2)).toThrow(
-                `The value random number should be a number (0 <= rand < 1)`,
+                'The value random number should be a number (0 <= rand < 1)',
             );
         });
 
         test('should throw an error if the list is not an array', () => {
             const list = {};
+
             expect(() => Helper.choiceInArray(list, 2)).toThrow(
                 'The list should be an array',
             );
