+++ packages/core/src/index.js
@@ -1,0 +1,5 @@
+import * as Game from './game.js';
+import * as Helper from './helper.js';
+import * as Shuffler from './shuffler.js';
+
+export { Game, Helper, Shuffler };
