+++ packages/core/src/shuffler.spec.js
